+++ packages/generator/src/inversify/models/serviceRule.ts
@@ -1,7 +1,9 @@
-import { LifeScope } from "../types";
+import { LifeScope, ServiceIdentifier } from "../types";
 
 export default interface ServiceRule {
   regexPattern: RegExp;
+  identifier?: ServiceIdentifier;
   scope: LifeScope;
   addDecorators?: boolean;
+  registerAutoFactory?: boolean;
 }